+++ pyplotformat/plot/__init__.py
@@ -3,4 +3,5 @@ This subpackage contains the plotting modules for the pyPlotFormat package.
 '''
 
 from .plot_2d import Format2D
+from .plot_polar import FormatPolar
 from .legend import FormatLegend